+++ test/renderers/class.test.js
@@ -1,0 +1,22 @@
+"use strict";
+
+describe("Renderer Class", () => {
+    it.skip("Create a default Renderer", () => {
+
+    });
+    it.skip("Extending Renderer class", () => {
+
+    });
+    it.skip("Output file", () => {
+
+    });
+    it.skip("Title", () => {
+
+    });
+    it.skip("Rendering multiple files", () => {
+
+    });
+    it.skip("Rendering in a directory",()=>{
+        
+    });
+});